mainly usercontent
also moved some directories around

+++ src/api/services/v1/storage/copy_overwrite.rs
@@ -5,7 +5,6 @@ use actix_web::{
 use mongodb::Database;
 use std::{
     error::Error,
-    ffi::OsStr,
     path::{Path as FilePath, PathBuf},
 };
 use tokio::{fs, io};
@@ -15,22 +14,23 @@ use goodmorning_bindings::{
     traits::ResTrait,
 };
 
-use crate::{functions::*, structs::*};
+use crate::{functions::*, structs::*, *};
 
-#[post("/copy")]
-pub async fn copy(
+#[post("/copy-overwrite")]
+pub async fn copy_overwrite(
     db: Data<Database>,
     post: Json<V1FromTo>,
     storage_limits: Data<StorageLimits>,
 ) -> Json<V1Response> {
     Json(V1Response::from_res(
-        copy_task(&post.from, &post.to, &post.token, &db, &storage_limits).await,
+        copy_overwrite_task(&post.from, &post.to, &post.from_userid, &post.token, &db, &storage_limits).await,
     ))
 }
 
-async fn copy_task(
+async fn copy_overwrite_task(
     from: &str,
     to: &str,
+    from_id: &str,
     token: &str,
     db: &Database,
     storage_limits: &StorageLimits,
@@ -45,39 +45,28 @@ async fn copy_task(
         }
     };
 
-    let path_buf = PathBuf::from(format!("usercontent/{}", account.id)).join(to);
-
-    if !editable(&path_buf) {
-        return Err(V1Error::NotEditable.into());
+    if !account.verified {
+        return Ok(V1Response::Error {
+            kind: V1Error::NotVerified,
+        });
     }
 
-    let from_buf = PathBuf::from(format!("usercontent/{}", from));
+    let to_buf = PathBuf::from(USERCONTENT.as_str()).join(&account.id).join(to.trim_start_matches('/'));
+    let from_buf = PathBuf::from(USERCONTENT.as_str()).join(from_id).join(from.trim_start_matches('/'));
 
-    if from_buf
-        .iter()
-        .any(|section| section == OsStr::new(".") || section == OsStr::new(".."))
-    {
-        return Err(V1Error::FileNotFound.into());
+    if !editable(&to_buf) || !has_dotdot(&to_buf) || !has_dotdot(&from_buf) {
+        return Err(V1Error::PermissionDenied.into());
     }
 
-    if !fs::try_exists(&path_buf).await? {
-        return Err(V1Error::FileNotFound.into());
-    }
-
-    let user = match from_buf.iter().nth(1) {
-        Some(id) => id.to_str().unwrap(),
-        None => return Err(V1Error::FileNotFound.into()),
-    };
-
-    if user != account.id
-        && (!fs::try_exists(&from_buf).await?
-            || Visibilities::visibility(&from_buf).await?.visibility == ItemVisibility::Private)
-    {
+    if from_id == account.id {
+        if !fs::try_exists(&from_buf).await? {
+            return Err(V1Error::FileNotFound.into());
+        }
+    } else if !fs::try_exists(&from_buf).await? || Visibilities::visibility(&from_buf).await?.visibility == ItemVisibility::Private {
         return Err(V1Error::FileNotFound.into());
     }
 
     let metadata = fs::metadata(&from_buf).await?;
-    println!("{:?}", path_buf);
 
     if metadata.is_file() {
         if account
@@ -86,8 +75,8 @@ async fn copy_task(
         {
             return Err(V1Error::FileTooLarge.into());
         }
-        remove(&path_buf).await?;
-        fs::copy(&from_buf, &path_buf).await?;
+        remove(&to_buf).await?;
+        fs::copy(&from_buf, &to_buf).await?;
     } else {
         if account
             .exceeds_limit(storage_limits, Some(dir_size(&from_buf).await?), None)
@@ -95,8 +84,8 @@ async fn copy_task(
         {
             return Err(V1Error::FileTooLarge.into());
         }
-        remove(&path_buf).await?;
-        copy_folder(&from_buf, &path_buf).await?;
+        remove(&to_buf).await?;
+        copy_folder(&from_buf, &to_buf).await?;
     }
 
     let mut from_visibilities = Visibilities::read_dir(from_buf.parent().unwrap()).await?;
@@ -112,20 +101,20 @@ async fn copy_task(
         }
     };
 
-    let file_name = path_buf
+    let file_name = to_buf
         .file_name()
         .unwrap()
         .to_os_string()
         .into_string()
         .unwrap();
 
-    if from_buf.parent() == path_buf.parent() {
+    if from_buf.parent() == to_buf.parent() {
         let _ = from_visibilities.0.insert(file_name, from_visibility);
-        from_visibilities.save(path_buf.parent().unwrap()).await?;
+        from_visibilities.save(to_buf.parent().unwrap()).await?;
     } else {
-        let mut new_visibilities = Visibilities::read_dir(path_buf.parent().unwrap()).await?;
+        let mut new_visibilities = Visibilities::read_dir(to_buf.parent().unwrap()).await?;
         new_visibilities.0.insert(file_name, from_visibility);
-        new_visibilities.save(path_buf.parent().unwrap()).await?;
+        new_visibilities.save(to_buf.parent().unwrap()).await?;
     }
 
     Ok(V1Response::Copied {
@@ -134,6 +123,9 @@ async fn copy_task(
 }
 
 async fn remove(path: &FilePath) -> io::Result<()> {
+    if fs::try_exists(path).await? {
+        return Ok(());
+    }
     if fs::metadata(&path).await?.is_dir() {
         fs::remove_dir_all(&path).await
     } else {
